requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2023.5.7
charset-normalizer==2.0.12
colorama==0.4.6
Django==3.2.18
django-bootstrap4==22.3
django-environ==0.9.0
django-multiselectfield==0.1.12
gunicorn==20.1.0
iamport-rest-client==0.9.0
idna==3.4
importlib-metadata==2.1.3
Pillow==9.4.0
pypng==0.20220715.0
pytz==2022.7.1
qrcode==7.4.2
requests==2.26.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.6
zipp==3.14.0
